backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
bcrypt
PyJWT
python-multipart